Fix a bug in the extraction of xforms.h from the specification.
The siftxfrms script pulls the xforms.h table directly from the
online brotli specification.  There was a bug where the HTML for
">", which is "&gt;", was not converted to ">".  This fixes that
bug in the script and updates xforms.h.  The script was also
updated to produce exactly the xforms.h file.  yeast.c has an
added trace which was used to find this problem.  check.c was
updated to remove extraneous output, and rerun to generate the
correct length and CRC for the table.  The brotli-04 specification
needs to be updated with the corrected values.