Update .br format to improve merging and clarify recommendations.
This moves the uncompressed length before to check value to allow
finding the check value immediately before the next header or the
final trailer.  This adds examples of the recommended uses of the
format for seekable media, in order to facilitate merging of .br
streams.  The notation of the structure is improved to distinguish
repeated from optional items, and to clarify the validity of a .br
stream with just a signature and a trailer.  The offset to the
previous header is moved to the beginning of the header to aid in
rapid scanning of a .br file.  An example of a valid stream with
just a signature and a trailer is added to the end.